requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file base.txt base.in
#
certifi==2018.8.24        # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
defusedxml==0.5.0         # via python3-openid
django-allauth==0.37.1
django-environ==0.4.5
django-rest-auth==0.9.3
django==2.1.1
djangorestframework==3.8.2
drfapikey==0.0.3
first==2.0.1              # via pip-tools
idna==2.7                 # via requests
oauthlib==2.1.0           # via requests-oauthlib
pip-tools==2.0.2
psycopg2-binary==2.7.5
python3-openid==3.1.0     # via django-allauth
pytz==2018.5              # via django
requests-oauthlib==1.0.0  # via django-allauth
requests==2.19.1          # via django-allauth, requests-oauthlib
six==1.11.0               # via django-rest-auth, pip-tools
urllib3==1.23             # via requests
